Refactor live vehicle tag
