--- src/components/Map/LiveVehicleMarker/styles.scss
@@ -1,51 +1,0 @@
-@import '../../../assets/styles/import.scss';
-
-.wrapper_outer {
-    border-radius: 50%;
-    height: 2.5rem;
-    width: 2.5rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
-    transition: all 0.2s ease-in-out;
-    &:hover {
-        transform: scale(1.5);
-    }
-}
-
-.wrapper_inner {
-    height: 90%;
-    width: 90%;
-    border-radius: 50%;
-    margin: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: $font-sizes-medium;
-}
-
-.ttc-wrapper {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: left;
-    padding-top: 0.2rem;
-}
-
-.first-line {
-    display: flex;
-    align-items: center;
-    margin-bottom: 0.4rem;
-    width: 100%;
-}
-
-.second-line {
-    white-space: nowrap;
-}
-
-.front-text {
-    margin-left: 0.8rem;
-    font-size: $font-sizes-large;
-    white-space: nowrap;
-}
